requirements.txt
Django==1.6.2
Pillow==2.3.0
boto==2.27.0
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-bootstrapped==1.6.4
django-ckeditor-updated==4.2.7
django-googlesearch==0.1.2
django-location-field==1.3.2
django-model-utils==2.0.2
django-redis==3.5.2
django-redis-cache==0.11.1
django-robots==1.0
django-storages==1.1.8
django-suit==0.2.7
gunicorn==18.0
hiredis==0.1.2
psycopg2==2.5.1
pystache==0.5.3
redis==2.9.1
sorl-thumbnail==11.12
static==0.4
wsgiref==0.1.2
